Moved conversions array to its own file.
